requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
anthropic>=0.7.0
python-dotenv==1.0.0
flask-migrate==4.0.5
pillow==10.0.1
werkzeug==2.3.7
email-validator==2.0.0.post2
